requirements.txt
openai
ollama
gradio
python-dotenv
requests